Contribution from Georg Fischer, wraptr.2021-10-17.19.tgz

+++ src/irvine/oeis/a029/A029614.java
@@ -8,8 +8,8 @@ import irvine.math.z.Z;
  */
 public class A029614 extends A029600 {
 
-  private long mN = 0;
-  private long mM = 0;
+  private int mN = 0;
+  private int mM = 0;
 
   @Override
   public Z next() {
