+++ src/irvine/oeis/a029/A029632.java
@@ -8,8 +8,8 @@ import irvine.math.z.Z;
  */
 public class A029632 extends A029618 {
 
-  private long mN = 0;
-  private long mM = 0;
+  private int mN = 0;
+  private int mM = 0;
 
   @Override
   public Z next() {
